requirements.txt
azure-functions
pytest
httpx
python-dotenv
